# ARROW-4725: [C++] Enable dictionary builder tests with MinGW build