@wesm 2019-05-09
@pitrou could you have a look at this if possible? Are there any performance implications?