@wesm 2019-05-09
@kou can you or someone run a benchmark comparison (we are working on doing this in an automated fashion in the near future)?