@pitrou 2019-05-09
I don't like this at all. This is adding some complication and exposing implementation details in the details just to appease an exotic platform compiler.

I'd also like to see some benchmark numbers.