@wesm 2019-05-09
I've got some refactoring of this file going in my branch for ARROW-3144 -- https://github.com/wesm/arrow/tree/ARROW-3144. It seems the issue is around symbol visibility of the template exports -- the solution may be to move more of the implementation to the header file. I could look at this in more detail if you'd like, but it won't be until next week

I don't think that exposing implementation in the header is a bad thing so long as the details are in an `internal` namespace or similar