@kou 2019-05-09
I didn't run `make benchmark`.

Now, I run it on my Linux desktop. Note that this result may have error because I was doing some light tasks such as writing a comment on the same machine while `make benchmark`.

Before:

```text
BM_BuildInt64DictionaryArrayRandom/repeats:2                     79 us         79 us       8750   960.194MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2                     80 us         80 us       8750   956.713MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_mean                80 us         80 us       8750   958.454MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_median              80 us         80 us       8750   958.454MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_stddev               0 us          0 us       8750   2.46103MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2                 74 us         74 us       9448   1033.77MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2                 74 us         74 us       9448   1026.17MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_mean            74 us         74 us       9448   1029.97MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_median          74 us         74 us       9448   1029.97MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_stddev           0 us          0 us       9448   5.37446MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2                    73 us         73 us       9558   1040.16MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2                    74 us         74 us       9558   1033.28MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_mean               74 us         74 us       9558   1036.72MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_median             74 us         74 us       9558   1036.72MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_stddev              0 us          0 us       9558   4.87107MB/s
BM_BuildStringDictionaryArray/repeats:2                         177 us        177 us       3965   565.818MB/s
BM_BuildStringDictionaryArray/repeats:2                         177 us        177 us       3965   564.514MB/s
BM_BuildStringDictionaryArray/repeats:2_mean                    177 us        177 us       3965   565.166MB/s
BM_BuildStringDictionaryArray/repeats:2_median                  177 us        177 us       3965   565.166MB/s
BM_BuildStringDictionaryArray/repeats:2_stddev                    0 us          0 us       3965   944.721kB/s
```

After:

```text
BM_BuildInt64DictionaryArrayRandom/repeats:2                     63 us         63 us      11060   1.19024GB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2                     63 us         63 us      11060   1.18231GB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_mean                63 us         63 us      11060   1.18627GB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_median              63 us         63 us      11060   1.18627GB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_stddev               0 us          0 us      11060   5.74351MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2                 56 us         56 us      12392   1.32746GB/s
BM_BuildInt64DictionaryArraySequential/repeats:2                 56 us         56 us      12392   1.32213GB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_mean            56 us         56 us      12392    1.3248GB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_median          56 us         56 us      12392    1.3248GB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_stddev           0 us          0 us      12392    3.8592MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2                    57 us         57 us      11946   1.31269GB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2                    56 us         56 us      11946   1.32086GB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_mean               57 us         57 us      11946   1.31677GB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_median             57 us         57 us      11946   1.31677GB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_stddev              0 us          0 us      11946   5.91462MB/s
BM_BuildStringDictionaryArray/repeats:2                         168 us        168 us       4162    593.52MB/s
BM_BuildStringDictionaryArray/repeats:2                         169 us        169 us       4162   592.553MB/s
BM_BuildStringDictionaryArray/repeats:2_mean                    169 us        169 us       4162   593.036MB/s
BM_BuildStringDictionaryArray/repeats:2_median                  169 us        169 us       4162   593.036MB/s
BM_BuildStringDictionaryArray/repeats:2_stddev                    0 us          0 us       4162   700.711kB/s
```

It seems that there are no significant slowdown. Most of cases are faster a bit but it may be error.

Some other builder benchmarks are slower with this change. But I'm not sure this is related to this change or just error.

Before:

```text
BM_BuildPrimitiveArrayNoNulls/repeats:2                       46565 us      46565 us         15   10.7376GB/s
BM_BuildPrimitiveArrayNoNulls/repeats:2                       46523 us      46524 us         15   10.7471GB/s
BM_BuildPrimitiveArrayNoNulls/repeats:2_mean                  46544 us      46545 us         15   10.7424GB/s
BM_BuildPrimitiveArrayNoNulls/repeats:2_median                46544 us      46545 us         15   10.7424GB/s
BM_BuildPrimitiveArrayNoNulls/repeats:2_stddev                   29 us         29 us         15   6.84988MB/s
```

After:

```text
BM_BuildPrimitiveArrayNoNulls/repeats:2                       91208 us      91209 us          8    5.4819GB/s
BM_BuildPrimitiveArrayNoNulls/repeats:2                       90612 us      90613 us          8   5.51799GB/s
BM_BuildPrimitiveArrayNoNulls/repeats:2_mean                  90910 us      90911 us          8   5.49994GB/s
BM_BuildPrimitiveArrayNoNulls/repeats:2_median                90910 us      90911 us          8   5.49994GB/s
BM_BuildPrimitiveArrayNoNulls/repeats:2_stddev                  421 us        422 us          8   26.1289MB/s
```

See the attached full benchmark logs for details:

* Before: [benchmark-logs.master.tar.gz](https://github.com/apache/arrow/files/3164034/benchmark-logs.master.tar.gz)
* After: [benchmark-logs.branch.tar.gz](https://github.com/apache/arrow/files/3164035/benchmark-logs.branch.tar.gz)