@wesm 2019-05-11
I'll look at the benchmark results also locally. 

@kou what is the easiest way to reproduce this issue locally, using Cygwin? I would like to understand the template visibility issue more deeply since I've spent a lot of time on such problems over the last few years