@wesm 2019-05-11
Hm this branch seems consistently faster than master

before

```
 ./release/arrow-builder-benchmark --benchmark_filter=Dict
2019-05-10 21:07:59
Running ./release/arrow-builder-benchmark
Run on (6 X 4400 MHz CPU s)
CPU Caches:
  L1 Data 32K (x6)
  L1 Instruction 32K (x6)
  L2 Unified 256K (x6)
  L3 Unified 12288K (x1)
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
-----------------------------------------------------------------------------------------------
Benchmark                                                        Time           CPU Iterations
-----------------------------------------------------------------------------------------------
BM_BuildInt64DictionaryArrayRandom/repeats:2                   121 us        121 us       6184   632.767MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2                   129 us        129 us       6184   593.177MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_mean              125 us        125 us       6184   612.972MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_median            125 us        125 us       6184   612.972MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_stddev              6 us          6 us       6184   27.9946MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2               101 us        101 us       7143   756.246MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2               102 us        102 us       7143   748.023MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_mean          101 us        101 us       7143   752.134MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_median        101 us        101 us       7143   752.134MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_stddev          1 us          1 us       7143   5.81457MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2                  102 us        102 us       6941    749.24MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2                  101 us        101 us       6941   752.723MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_mean             102 us        102 us       6941   750.981MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_median           102 us        102 us       6941   750.981MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_stddev             0 us          0 us       6941   2.46331MB/s
BM_BuildStringDictionaryArray/repeats:2                        230 us        230 us       2987   434.809MB/s
BM_BuildStringDictionaryArray/repeats:2                        231 us        231 us       2987   432.284MB/s
BM_BuildStringDictionaryArray/repeats:2_mean                   231 us        231 us       2987   433.546MB/s
BM_BuildStringDictionaryArray/repeats:2_median                 231 us        231 us       2987   433.546MB/s
BM_BuildStringDictionaryArray/repeats:2_stddev                   1 us          1 us       2987   1.78519MB/s
```

after

```
$ ./release/arrow-builder-benchmark --benchmark_filter=Dict
2019-05-10 21:08:21
Running ./release/arrow-builder-benchmark
Run on (6 X 4400 MHz CPU s)
CPU Caches:
  L1 Data 32K (x6)
  L1 Instruction 32K (x6)
  L2 Unified 256K (x6)
  L3 Unified 12288K (x1)
***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
-----------------------------------------------------------------------------------------------
Benchmark                                                        Time           CPU Iterations
-----------------------------------------------------------------------------------------------
BM_BuildInt64DictionaryArrayRandom/repeats:2                    84 us         83 us       9113   920.758MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2                    84 us         83 us       9113   918.057MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_mean               84 us         83 us       9113   919.407MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_median             84 us         83 us       9113   919.407MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_stddev              0 us          0 us       9113   1.91017MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2                75 us         74 us       9367   1028.33MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2                75 us         74 us       9367   1025.62MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_mean           75 us         74 us       9367   1026.98MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_median         75 us         74 us       9367   1026.98MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_stddev          0 us          0 us       9367    1.9155MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2                   73 us         73 us      10610   1044.79MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2                   72 us         72 us      10610   1059.06MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_mean              73 us         73 us      10610   1051.92MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_median            73 us         73 us      10610   1051.92MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_stddev             1 us          1 us      10610   10.0899MB/s
BM_BuildStringDictionaryArray/repeats:2                        210 us        210 us       3325   475.559MB/s
BM_BuildStringDictionaryArray/repeats:2                        210 us        210 us       3325   476.206MB/s
BM_BuildStringDictionaryArray/repeats:2_mean                   210 us        210 us       3325   475.882MB/s
BM_BuildStringDictionaryArray/repeats:2_median                 210 us        210 us       3325   475.882MB/s
BM_BuildStringDictionaryArray/repeats:2_stddev                   0 us          0 us       3325   468.806kB/s
```

@pitrou thoughts? I fear that may be due for multiple rounds of refactoring, we may also want to expand the benchmarking to be more comprehensive or understand why this patch makes things a good deal faster